website/requirements.txt
asgiref==3.7.2
Django==4.2.8
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
asgiref==3.7.2
Django==4.2.8
djangorestframework==3.14.0
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
